refactor: deferredJob module
