+++ src/lib/model/lib/index.js
@@ -1,0 +1,9 @@
+const deferredJob = require('./deferredJob');
+const request = require('./requests');
+const shared = require('./shared');
+
+module.exports = {
+    deferredJob,
+    request,
+    shared,
+};
