+++ src/lib/model/lib/shared/index.js
@@ -439,8 +439,6 @@ const mojaloopBulkPrepareToInternalBulkTransfer = (external, bulkQuotes, ilp) =>
     return internal;
 };
 
-const deferredJob = require('./deferredJob');
-
 module.exports = {
     internalPartyToMojaloopParty,
     internalQuoteResponseToMojaloop,
@@ -452,6 +450,5 @@ module.exports = {
     mojaloopTransactionRequestToInternal,
     mojaloopBulkQuotesRequestToInternal,
     internalBulkQuotesResponseToMojaloop,
-    mojaloopBulkPrepareToInternalBulkTransfer,
-    deferredJob
+    mojaloopBulkPrepareToInternalBulkTransfer
 };
